添加分类

+++ app/src/main/java/com/example/kotlin/viewholder/QAViewHolder.kt
@@ -1,13 +1,10 @@
 package com.example.kotlin.viewholder
 
 import android.graphics.Color
-import android.os.Build
 import android.view.View
-import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.RecyclerView
 import com.example.kotlin.data.QAItemData
 import kotlinx.android.synthetic.main.list_items.view.*
-import kotlinx.android.synthetic.main.list_items.view.list_item_text
 import kotlinx.android.synthetic.main.q_a_list_items.view.*
 
 /**
@@ -20,50 +17,58 @@ class QAViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     var currentPosition: Int = 0
     private lateinit var listener: HolderListener
 
-    init {
-        itemView.setOnClickListener {
-            listener.jump()
-        }
-        itemView.q_a_list_item_image.setOnClickListener { listener.delete() }
-        itemView.q_a_list_item_image.setOnLongClickListener {
-            listener.longClick()
-            return@setOnLongClickListener true
-        }
-    }
-
 
     fun setData(data: QAItemData, pos: Int) {
-        itemView.q_a_list_item_text.text = data?.title
-        data?.image?.let { itemView.q_a_list_item_image.setBackgroundResource(it) }
-        when {
-            data.tag == 1 -> {
-                itemView.q_a_list_item_text.setTextColor(Color.parseColor("#FF0000"))
+        if (data.head != 0) {
+            itemView.card_layout.visibility = View.GONE
+            itemView.content_head.visibility = View.VISIBLE
+            itemView.content_head.text = data.title
+        } else {
+            itemView.setOnClickListener {
+                listener.jump()
             }
-            data.tag == 2 -> {
-                itemView.q_a_list_item_text.setTextColor(Color.parseColor("#FFC1C1"))
+            itemView.q_a_list_item_image.setOnClickListener { listener.delete() }
+            itemView.q_a_list_item_image.setOnLongClickListener {
+                listener.longClick()
+                return@setOnLongClickListener true
             }
-            data.tag == 3 -> {
-                itemView.q_a_list_item_text.setTextColor(Color.parseColor("#FFD700"))
+
+
+
+            itemView.card_layout.visibility = View.VISIBLE
+            itemView.content_head.visibility = View.GONE
+            itemView.q_a_list_item_text.text = data?.title
+            data?.image?.let { itemView.q_a_list_item_image.setBackgroundResource(it) }
+            when {
+                data.tag == 1 -> {
+                    itemView.q_a_list_item_text.setTextColor(Color.parseColor("#FF0000"))
+                }
+                data.tag == 2 -> {
+                    itemView.q_a_list_item_text.setTextColor(Color.parseColor("#FFC1C1"))
+                }
+                data.tag == 3 -> {
+                    itemView.q_a_list_item_text.setTextColor(Color.parseColor("#FFD700"))
+                }
+                data.tag == 4 -> {
+                    itemView.q_a_list_item_text.setTextColor(Color.parseColor("#006400"))
+                }
+                data.tag > 4 -> {
+                    itemView.q_a_list_item_text.setTextColor(Color.parseColor("#00FF00"))
+                }
             }
-            data.tag == 4 -> {
-                itemView.q_a_list_item_text.setTextColor(Color.parseColor("#006400"))
+            if (data.lastTime != null && data.lastTime.isNotEmpty()) {
+                itemView.remember_last_time.text = "last : " + data.lastTime
+            } else {
+                itemView.remember_last_time.text = "未学习"
             }
-            data.tag > 4 -> {
-                itemView.q_a_list_item_text.setTextColor(Color.parseColor("#00FF00"))
+            if (data.nextTime != null && data.nextTime.isNotEmpty()) {
+                itemView.remember_next_time.text = "next : " + data.nextTime
+            } else {
+                itemView.remember_next_time.text = ""
             }
+            this.currentData = data
+            this.currentPosition = pos
         }
-        if(data.lastTime!=null&&data.lastTime.isNotEmpty()){
-            itemView.remember_last_time.text = "last : "+data.lastTime
-        }else{
-            itemView.remember_last_time.text = "未学习"
-        }
-        if(data.nextTime!=null&&data.nextTime.isNotEmpty()){
-            itemView.remember_next_time.text = "next : "+data.nextTime
-        }else{
-            itemView.remember_next_time.text = ""
-        }
-        this.currentData = data
-        this.currentPosition = pos
     }
 
     fun setOnListener(deleteListener: HolderListener) {
